chore(webclient): add repeating error report; chore(webclient): remove js deadcode.
